InsuranceContractsController.jl

InsuranceContractsController.Contract — Type

Contract

a component of a bitemporal entity

InsuranceContractsController.ContractPartnerRef — Type

ContractPartnerRef

a component of a bitemporal entity

InsuranceContractsController.ContractPartnerRefRevision — Type

ContractPartnerRefRevision

a revision of a Contract component of a bitemporal entity

InsuranceContractsController.ContractRevision — Type

ContractRevision

a revision of a Contract component of a bitemporal entity