InsuranceContractsController.jl

InsuranceContracts.Contract — Type

Contract

a contract component of a bitemporal entity

InsuranceContracts.ContractPartnerRef — Type

ContractPartnerRef

a partner reference of a contract component, i.e. policy holder, premium payer

InsuranceContracts.ContractPartnerRefRevision — Type

ContractPartnerRefRevision

a revision of a contract's partner reference

InsuranceContracts.ContractRevision — Type

ContractRevision

a revision of a contract component

InsuranceContracts.ProductItem — Type

ProductItem

a productitem component of a contract component

InsuranceContracts.ProductItemPartnerRef — Type

ProductItemPartnerRef

a reference to a partner of a productitem, i.e. insured person

InsuranceContracts.ProductItemPartnerRefRevision — Type

ProductItemPartnerRefRevision

a revision of a productItem's partner reference

InsuranceContracts.ProductItemRevision — Type

ProductItemRevision

a revision of a productitem component

InsuranceContracts.ProductItemTariffRef — Type

ProductItemTariffRef

a reference to a tariff of a productitem

InsuranceContracts.ProductItemTariffRefRevision — Type

ProductItemTariffRefRevision

a revision of a productitem's reference to a tariff

InsurancePartners.Partner — Type

Partner

a component of a bitemporal entity

InsurancePartners.PartnerRevision — Type

Partner_Revision

a revision of a Partner component of a bitemporal entity

InsuranceTariffs.Tariff — Type

Tariff

a component of a bitemporal entity

InsuranceTariffs.TariffRevision — Type

Tariff_Revision

a revision of a Tariff component of a bitemporal entity